Port upstream Pillow 10.0.0
This work is based off Pillow 10.0.0 configured on Linux x86_64 and
adapted to work with Unikraft.
Files under importfix/ are correctly namespaced wrapper modules for any
binary modules that we compile and link into Unikraft, and must be
copied over to the python rootfs.
Selecting this library will add its files to the python rootfs build.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>